# services/sql-assessment-service/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "SQL Assessment Service API",
    "version": "1.0.0",
    "description": "REST API for SQL exercise generation, student query grading, and natural-language description of SQL queries in an educational context.",
    "contact": {
      "name": "ALADIN Project"
    }
  },
...
  "paths": {
...
    "/api/description/template": {
      "post": {
        "summary": "Generate a template-based SQL description",
        "description": "Produces a deterministic, AST-driven natural-language description of the supplied SQL query using a hand-crafted template engine. No LLM or external API call is made.\n",
        "tags": [
          "Description"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DescriptionRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Description generated successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DescriptionResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request, unregistered database, or SQL parse failure.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Template engine failure.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/description/llm/default": {
      "post": {
        "summary": "Generate an LLM description (default, temperature 0)",
        "description": "Uses a single-shot LLM call at temperature 0 to produce a natural-language description of the SQL query. Requires OPENAI_API_KEY to be set; falls back to the template engine otherwise.\n",
        "tags": [
          "Description"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DescriptionRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Description generated successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DescriptionResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request or unregistered database.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "LLM call failure.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/description/llm/creative": {
      "post": {
        "summary": "Generate a creative LLM description (temperature 0.7)",
        "description": "Uses a single-shot LLM call at temperature 0.7 to produce a more expressive natural-language description of the SQL query.\n",
        "tags": [
          "Description"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DescriptionRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Description generated successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DescriptionResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request or unregistered database.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "LLM call failure.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/description/llm/multi-step": {
      "post": {
        "summary": "Generate a multi-step LLM description",
        "description": "Runs a three-stage chained LLM pipeline to progressively build a high-quality natural-language description of the SQL query.\n",
        "tags": [
          "Description"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DescriptionRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Description generated successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DescriptionResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request or unregistered database.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "LLM pipeline failure.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/description/hybrid": {
      "post": {
        "summary": "Generate a hybrid template + LLM description",
        "description": "Feeds the deterministic template engine output as context into an LLM for natural-language post-processing, combining the reliability of templates with the fluency of an LLM.\n",
        "tags": [
          "Description"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DescriptionRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Description generated successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DescriptionResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request, unregistered database, or SQL parse failure.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Hybrid generation failure.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "ErrorResponse": {
        "type": "object",
        "required": [
          "message"
        ],
        "properties": {
          "message": {
            "type": "string",
            "description": "Human-readable error description (localised).",
            "example": "Unable to connect to the database."
          },
          "code": {
            "type": "string",
            "description": "Optional machine-readable error code.",
            "example": "NON_SELECT"
          }
        }
      },
...
      "PostgresConnectionInfo": {
        "type": "object",
        "required": [
          "type",
          "host",
          "port",
          "username",
          "password",
          "database",
          "schema"
        ],
        "properties": {
          "type": {
            "type": "string",
            "enum": [
              "postgres"
            ],
            "example": "postgres"
          },
          "host": {
            "type": "string",
            "example": "localhost"
          },
          "port": {
            "type": "number",
            "example": 5432
          },
          "username": {
            "type": "string",
            "example": "myuser"
          },
          "password": {
            "type": "string",
            "example": "mypassword"
          },
          "database": {
            "type": "string",
            "example": "mydb"
          },
          "schema": {
            "type": "string",
            "example": "public"
          }
        }
      },
...
      "DescriptionRequest": {
        "type": "object",
        "required": [
          "connectionInfo",
          "query"
        ],
        "properties": {
          "connectionInfo": {
            "$ref": "#/components/schemas/PostgresConnectionInfo"
          },
          "query": {
            "type": "string",
            "description": "Raw SQL query string to describe.",
            "example": "SELECT name FROM customers WHERE age > 30"
          },
          "isSelfJoin": {
            "type": "boolean",
            "description": "Whether the query involves a self-join. Defaults to false."
          },
          "languageCode": {
            "type": "string",
            "description": "BCP 47 language code for the generated description. Defaults to \"en\".",
            "example": "en"
          }
        }
      },
      "DescriptionResponse": {
        "type": "object",
        "required": [
          "description",
          "languageCode"
        ],
        "properties": {
          "description": {
            "type": "string",
            "description": "Natural-language description of the SQL query.",
            "example": "Retrieve the names of all customers older than 30."
          },
          "languageCode": {
            "type": "string",
            "description": "The language code of the returned description.",
            "example": "en"
          }
        }
      },
...
    }
  }
}